ensure installed, move stuff (wip)
