+++ src/commands/listRemote.js
@@ -1,5 +1,5 @@
-const log = require('../common/log')
-const { getVersionsFromTags, printVersions } = require('../common/utils')
+const log = require('../util/log')
+const { getVersionsFromTags, printVersions } = require('../util/utils')
 
 const listRemoteCommand = () => {
     log('list-remote')
