+++ src/util/utils.js
@@ -1,8 +1,10 @@
+const fs = require('fs')
 const os = require('os')
 const path = require('path')
 const request = require('request')
 
-const log = require('../common/log')
+const installVersion = require('../commands/install')
+const log = require('./log')
 
 const yvmPath = process.env.YVM_DIR || path.resolve(os.homedir(), '.yvm')
 const versionRootPath = rootPath => path.resolve(rootPath, 'versions')
@@ -44,7 +46,16 @@ const getVersionsFromTags = () => {
     })
 }
 
+const ensureVersionInstalled = (version, rootPath = yvmPath) => {
+    const yarnBinDir = getExtractionPath(version, rootPath)
+    if (fs.existsSync(yarnBinDir)) {
+        return Promise.resolve()
+    }
+    return installVersion(version, rootPath)
+}
+
 module.exports = {
+    ensureVersionInstalled,
     getExtractionPath,
     getVersionsFromTags,
     printVersions,
